Full trip filter and already in filter applied when showing all trips.  User gotten right after login.
User gotten when logging in now instead of at some point after.
